Move code to node core, will make MmrQueryController a meta requirement
