+++ packages/node-core/src/subcommands/index.ts
@@ -2,3 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './forceClean.service';
+export * from './forceClean.module';
+export * from './mmrRegnerate.module';
+export * from './mmrMigrate.module';
+export * from './reindex.service';
