+++ packages/node/src/meta/meta.module.ts
@@ -2,15 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Module } from '@nestjs/common';
-import {
-  MetricEventListener,
-  ReadyController,
-  ReadyService,
-  HealthController,
-  HealthService,
-  MmrQueryController,
-  gaugeProviders,
-} from '@subql/node-core';
+import { metaServices, metaControllers } from '@subql/node-core';
 import { PrometheusModule } from '@willsoto/nestjs-prometheus';
 import { FetchModule } from '../indexer/fetch.module';
 import { MetaController } from './meta.controller';
@@ -18,18 +10,7 @@ import { MetaService } from './meta.service';
 
 @Module({
   imports: [PrometheusModule.register(), FetchModule],
-  controllers: [
-    MetaController,
-    HealthController,
-    ReadyController,
-    MmrQueryController,
-  ],
-  providers: [
-    MetricEventListener,
-    ...gaugeProviders,
-    MetaService,
-    HealthService,
-    ReadyService,
-  ],
+  controllers: [...metaControllers, MetaController],
+  providers: [...metaServices, MetaService],
 })
 export class MetaModule {}
