+++ packages/node/src/subcommands/forceClean.module.ts
@@ -2,15 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Module } from '@nestjs/common';
-import { SchedulerRegistry } from '@nestjs/schedule';
-import { DbModule, ForceCleanService } from '@subql/node-core';
+import { DbModule, ForceCleanFeatureModule } from '@subql/node-core';
 import { ConfigureModule } from '../configure/configure.module';
-
-@Module({
-  providers: [ForceCleanService, SchedulerRegistry],
-  controllers: [],
-})
-export class ForceCleanFeatureModule {}
 
 @Module({
   imports: [
